requirements.txt
# Core analysis dependencies (Tier 1)
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
statsmodels>=0.14
scipy>=1.10
scikit-learn>=1.3
pyarrow>=12.0
h5py>=3.8

# Tier 2: raw data ingest (optional — only needed for rebuild from source)
requests>=2.28
python-dotenv>=1.0
